requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
PyMySQL==1.1.0
Pillow==10.1.0
python-dotenv==1.0.0
